Rename package and cleanup code
