+++ PluginCore/src/main/java/com/mooner/starlight/plugincore/config/data/ConfigCategoryImpl.kt
@@ -1,4 +1,12 @@
-package com.mooner.starlight.plugincore.config
+/*
+ * ConfigCategoryImpl.kt created by Minki Moon(mooner1022)
+ * Copyright (c) mooner1022. all rights reserved.
+ * This code is licensed under the GNU General Public License v3.0.
+ */
+
+package com.mooner.starlight.plugincore.config.data
+
+import com.mooner.starlight.plugincore.config.TypedString
 
 class ConfigCategoryImpl(
     private val data: Map<String, TypedString>
