+++ PluginCore/src/main/java/com/mooner/starlight/plugincore/config/data/FileConfig.kt
@@ -1,6 +1,13 @@
-package com.mooner.starlight.plugincore.config
+/*
+ * FileConfig.kt created by Minki Moon(mooner1022)
+ * Copyright (c) mooner1022. all rights reserved.
+ * This code is licensed under the GNU General Public License v3.0.
+ */
+
+package com.mooner.starlight.plugincore.config.data
 
 import com.mooner.starlight.plugincore.Session.json
+import com.mooner.starlight.plugincore.config.TypedString
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
