+++ PluginCore/src/main/java/com/mooner/starlight/plugincore/config/data/GlobalConfig.kt
@@ -1,7 +1,15 @@
-package com.mooner.starlight.plugincore.config
+/*
+ * GlobalConfig.kt created by Minki Moon(mooner1022)
+ * Copyright (c) mooner1022. all rights reserved.
+ * This code is licensed under the GNU General Public License v3.0.
+ */
+
+package com.mooner.starlight.plugincore.config.data
 
 import android.view.View
 import com.mooner.starlight.plugincore.Session.json
+import com.mooner.starlight.plugincore.config.TypedString
+import com.mooner.starlight.plugincore.config.typed
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
