+++ PluginCore/src/main/java/com/mooner/starlight/plugincore/config/data/MutableConfigCategory.kt
@@ -1,4 +1,12 @@
-package com.mooner.starlight.plugincore.config
+/*
+ * MutableConfigCategory.kt created by Minki Moon(mooner1022)
+ * Copyright (c) mooner1022. all rights reserved.
+ * This code is licensed under the GNU General Public License v3.0.
+ */
+
+package com.mooner.starlight.plugincore.config.data
+
+import com.mooner.starlight.plugincore.config.TypedString
 
 class MutableConfigCategory(
     val data: MutableMap<String, TypedString>
