+++ app/src/main/res/layout/config_button_flat.xml
@@ -1,10 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-                                                   xmlns:app="http://schemas.android.com/apk/res-auto"
-                                                   android:id="@+id/layout_configButton"
-                                                   android:layout_width="match_parent"
-                                                   android:layout_height="wrap_content" android:paddingTop="12dp"
-                                                   android:paddingBottom="12dp" android:minHeight="64dp">
+<androidx.constraintlayout.widget.ConstraintLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/layout_configButton"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingTop="12dp"
+        android:paddingBottom="12dp"
+        android:minHeight="64dp">
 
     <ImageView
             android:id="@+id/icon"
@@ -12,7 +15,7 @@
             android:layout_height="24dp"
             android:layout_marginStart="16dp"
             android:src="@drawable/ic_round_keyboard_arrow_right_24"
-            android:tint="#000000"
+            android:tint="@color/text"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
@@ -28,16 +31,22 @@
             android:ellipsize="end"
             android:maxLines="1"
             android:textSize="16sp"
+            android:layout_marginEnd="8dp"
             app:layout_constraintStart_toEndOf="@+id/icon"
-            app:layout_constraintTop_toTopOf="parent" app:layout_constraintBottom_toTopOf="@+id/description"
-            app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"/>
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toTopOf="@+id/description"
+            app:layout_constraintEnd_toEndOf="parent" android:textColor="@color/text"/>
     <TextView
             android:text="Description"
             android:layout_width="0dp"
-            android:layout_height="wrap_content" android:id="@+id/description"
-            app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintStart_toEndOf="@+id/icon"
-            android:layout_marginStart="8dp" android:visibility="visible"
-            android:fontFamily="@font/nanumsquare_round_light" app:layout_constraintEnd_toEndOf="parent"
-            android:layout_marginEnd="8dp"/>
+            android:layout_height="wrap_content"
+            android:id="@+id/description"
+            android:layout_marginStart="8dp"
+            android:visibility="visible"
+            android:fontFamily="@font/nanumsquare_round_light"
+            android:layout_marginEnd="8dp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/icon"
+            app:layout_constraintEnd_toEndOf="parent" android:textColor="@color/text"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
