+++ app/src/main/res/layout/config_toggle.xml
@@ -13,6 +13,7 @@
             android:layout_width="24dp"
             android:layout_height="24dp"
             android:layout_marginStart="16dp"
+            android:tint="@color/text"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
@@ -38,9 +39,10 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content" android:id="@+id/description"
             app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintStart_toEndOf="@+id/icon"
-            android:layout_marginStart="8dp" android:visibility="gone"
-            android:fontFamily="@font/nanumsquare_round_light" android:layout_marginBottom="8dp"
-            app:layout_constraintEnd_toStartOf="@+id/toggle" android:layout_marginEnd="8dp"/>
+            android:layout_marginStart="8dp" android:visibility="visible"
+            android:fontFamily="@font/nanumsquare_round_light"
+            app:layout_constraintEnd_toStartOf="@+id/toggle" android:layout_marginEnd="8dp"
+            android:layout_marginBottom="8dp"/>
     <com.google.android.material.switchmaterial.SwitchMaterial
             android:id="@+id/toggle"
             android:layout_width="wrap_content"
@@ -53,5 +55,5 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            tools:ignore="UseSwitchCompatOrMaterialXml"/>
+            tools:ignore="UseSwitchCompatOrMaterialXml" android:minWidth="48dp" android:minHeight="48dp"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
