+++ app/src/main/res/layout/fragment_projects.xml
@@ -74,10 +74,12 @@
                 android:foregroundGravity="center"
                 android:padding="16dp"
                 android:src="@drawable/ic_round_add_white_24"
-                app:tint="@color/white" app:elevation="0dp"
-                app:rippleColor="@color/main_purple" app:backgroundTint="@color/main_purple"
                 android:layout_marginStart="8dp"
-                android:layout_marginEnd="@dimen/card_margin_horizontal"/>
+                android:layout_marginEnd="@dimen/card_margin_horizontal"
+                app:tint="@color/white"
+                app:elevation="0dp"
+                app:rippleColor="@color/main_purple"
+                app:backgroundTint="@color/main_purple"/>
     </LinearLayout>
 
     <androidx.cardview.widget.CardView
